V0.3. Bug Fixes. Matlab App and Standalone App
bug fix: if the user selects no inputs and then clicks generate the uialert pops up correctly but the function still executes the else statement. Return added to fix this.
Yellow Hex color seems to get lost in the matlab app designer when reopened - no known fix, but everything else in the layout is correct.