Fixed bugs, updated output to match hexapod input gui
Added splines to start up and stop of multisine generation - ensures smooth hexapod operation during an experiment.
Added scaling of maximum amplitude for translations and rotations, currently not a user input - will be in the release.
Added new export feature to export in a way that the hexapod gui can directly interface with
added info section to the data.mat file that contains all the meta data info
fixed bug in plotting that would sometimes select the wrong signal to plot in the multisine graph.