Version 0.3 of Multisine App
Compatibility with hexapod GUI fixed.
added slider for max amplitude
added smooth data head and tails to either side of multisine
changed export format to be more in line with testing methodology
Added info field to data structure that gives the number of period, realizations and all the DFT frequencies in the band selected
After pressing generate multisine the spin boxes for the min and max frequency update to what is actually being used.