V0.5 changes
Changed to a lcm sampling approach. app.Fs_U is the sampling rate for the hexapod. app.Fs_Y is the sampling rate of the measured force signal. we use the lcm between these two to generate our multisine signal. We then down sample data.u_out to app.Fs_U and data.u_measure and data.r to app.Fs_Y. This slows down the generation but we get matched signals for our end uses.