V0.6.1 Update to Multisine Generator App
Generator app now outputs the acceleration measurement in the U_data file.
Scaling is now done in the Amplitude Excited Harmonics section and is scaled for both linear and rotational acceleration.
Fixed bug in data.u and data.r where it was still possible to have old field names present.